03-async-job-processing/api/requirements.txt
flask
psycopg2-binary
redis
rq